update gateway_upgrade program
