+++ programs/examples/gateway_upgrade/src/instructions/mod.rs
@@ -1,0 +1,4 @@
+pub mod admin;
+pub mod deposit;
+pub mod execute;
+pub mod withdraw;
