+++ programs/examples/gateway_upgrade/src/utils/verify_and_update_nonce.rs
@@ -1,0 +1,17 @@
+use crate::errors::Errors;
+use crate::state::Pda;
+use anchor_lang::prelude::*;
+
+/// Verifies provided nonce is correct and updates pda nonce.
+pub fn verify_and_update_nonce(pda: &mut Account<Pda>, nonce: u64) -> Result<()> {
+    if nonce != pda.nonce {
+        msg!(
+            "Mismatch nonce: provided nonce = {}, expected nonce = {}",
+            nonce,
+            pda.nonce,
+        );
+        return err!(Errors::NonceMismatch);
+    }
+    pda.nonce += 1;
+    Ok(())
+}
